requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
PyMuPDF
beautifulsoup4
